Bench: 33465110

+++ Halogen/src/Network.h
@@ -39,11 +39,11 @@ public:
     static void Init();
 
 private:
-    std::vector<std::array<int16_t, HIDDEN_NEURONS>> Zeta;
+    std::vector<std::array<float, HIDDEN_NEURONS>> Zeta;
 
-    static std::array<std::array<int16_t, HIDDEN_NEURONS>, INPUT_NEURONS> hiddenWeights;
-    static std::array<int16_t, HIDDEN_NEURONS> hiddenBias;
-    static std::array<int16_t, HIDDEN_NEURONS> outputWeights;
-    static int16_t outputBias;
+    static std::array<std::array<float, HIDDEN_NEURONS>, INPUT_NEURONS> hiddenWeights;
+    static std::array<float, HIDDEN_NEURONS> hiddenBias;
+    static std::array<float, HIDDEN_NEURONS> outputWeights;
+    static float outputBias;
 };
